support sutrcutred data in add example dialog
