+++ app/web_ui/src/routes/(app)/generate/[project_id]/[task_id]/data_guide_setup/+page.svelte
@@ -5,7 +5,8 @@
   import { onMount, onDestroy } from "svelte"
   import { page } from "$app/stores"
   import { goto } from "$app/navigation"
-  import GuideSetupForm, { type GuideSample } from "./guide_setup_form.svelte"
+  import GuideSetupForm from "./guide_setup_form.svelte"
+  import type { GuideSample } from "$lib/components/add_example_dialog.svelte"
   import { pending_data_guide_example } from "./pending_example_store"
   import { get } from "svelte/store"
   import GuidePreview from "./guide_preview.svelte"
